requirements.txt
# FastHTML and dependencies
python-fasthtml
fastcore
uvicorn

# YAML support
pyyaml

# Optional: for production deployment
gunicorn
python-multipart
